requirements.txt
Flask==1.0.2
Pillow
google-cloud-storage